Move get_asset_info into trait
